+++ core-ng/src/main/java/core/framework/api/mongo/Mongo.java
@@ -1,12 +1,8 @@
 package core.framework.api.mongo;
 
-import org.bson.Document;
-
 /**
  * @author neo
  */
 public interface Mongo {
     void dropCollection(String collection);
-
-    Document eval(String function, Object... arguments);
 }